Ajustes hasta 11 del 6

+++ src/app/alumnos/page.js
@@ -1,6 +1,7 @@
 'use client';
-import Link from 'next/link';
 import React, { useEffect, useState } from 'react';
+import Link from 'next/link';
+import Swal from 'sweetalert2';
 
 export default function AlumnosPage(){
   const [users, setUsers] = useState([]);             // creamos el estado 'users'
@@ -10,7 +11,6 @@ export default function AlumnosPage(){
     fetch('/api/alumnos')                             // Ejecutamos FETCH
       .then((respuesta) => respuesta.json())          // Devuelve promesa, y retornamos .json()
       .then((users) => {
-        console.log('users',users);                   // Devuelve promesa y actualizamos users
         setUsers(users);
         setUsersInit(users);
       })
@@ -24,12 +24,39 @@ export default function AlumnosPage(){
   }
 
   const deleteStudentHandler = (id) => {
-    fetch('http://localhost:3000/api/alumnos', {
-      method: 'DELETE',
-      body: id
+    const student = users.filter(user => user._id === id);
+
+    Swal.fire({
+      title: "¿Estás seguro/a?",
+      text: `Estas por eliminar al alumno/a: ${student[0].nombre} ${student[0].apellido}`,
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonText: "Eliminar Alumno/a",
+      cancelButtonText: 'Cancelar'
+    }).then(result => {
+      if(result.isConfirmed){
+        fetch('http://localhost:3000/api/alumnos', {
+          method: 'DELETE',
+          body: id
+        })
+        .then(response => response.json())
+        .then(data => {
+          const { response } = data;
+          const usuariosFiltrados = users.filter( user => user._id !== response._id );
+          const usuariosFiltradosInit = usersInit.filter( user => user._id !== response._id );
+
+          setUsers(usuariosFiltrados);
+          setUsersInit(usuariosFiltradosInit);
+
+          Swal.fire({
+            title: 'Alumno eliminado',
+            text: 'El alumno fue eliminado correctamente',
+            icon: 'success',
+            confirmButtonText: 'Ok'
+          });
+        })
+      }
     })
-    .then(response => response.json())
-    .then(data => console.log(data))
   }
 
   return <div className='data-table'>
